+++ scripts/train_stable_sync.sh
@@ -93,6 +93,7 @@ PROJECT=${PROJECT:-mario-a3c}
 SAVE_ROOT=${SAVE_ROOT:-trained_models}
 LOG_ROOT=${LOG_ROOT:-tensorboard/a3c_super_mario_bros}
 DEVICE=${DEVICE:-auto}
+EPISODE_EVENTS_PATH=${EPISODE_EVENTS_PATH:-}
 
 RUN_SAVE_DIR="${SAVE_ROOT%/}/${RUN_NAME}"
 RUN_LOG_DIR="${LOG_ROOT%/}/${RUN_NAME}"
@@ -119,6 +120,7 @@ CMD=(
   --save-dir "$RUN_SAVE_DIR"
   --log-dir "$RUN_LOG_DIR"
   --metrics-path "$RUN_METRICS_PATH"
+  ${EPISODE_EVENTS_PATH:+--episodes-event-path "$EPISODE_EVENTS_PATH"}
   --project "$PROJECT"
   --device "$DEVICE"
   --reward-distance-weight "$REWARD_DISTANCE_WEIGHT"